File Storage Test: Added tests for file storage and refactored static file mounting
