+++ ellar/core/files/storages/utils.py
@@ -1,9 +1,8 @@
 import os
 import re
 import typing
-from io import BytesIO
 
-from .exceptions import SuspiciousFileOperation
+from .exceptions import UnsafeFileOperation
 
 
 def copy_stream(
@@ -19,11 +18,6 @@ def copy_stream(
     return length
 
 
-def get_length(source: typing.IO) -> int:
-    buffer = BytesIO()
-    return copy_stream(source, buffer)
-
-
 def get_valid_filename(name: str) -> str:
     """
     Return the given string converted to a string that can be used for a clean
@@ -36,16 +30,16 @@ def get_valid_filename(name: str) -> str:
     s = str(name).strip().replace(" ", "_")
     s = re.sub(r"(?u)[^-\w.]", "", s)
     if s in {"", ".", ".."}:
-        raise Exception("Could not derive file name from '%s'" % name)
+        raise UnsafeFileOperation("Could not derive file name from '%s'" % name)
     return s
 
 
 def validate_file_name(name: str) -> str:
     # Remove potentially dangerous names
     if os.path.basename(name) in {"", ".", ".."}:
-        raise SuspiciousFileOperation("Could not derive file name from '%s'" % name)
+        raise UnsafeFileOperation("Could not derive file name from '%s'" % name)
 
     if name != os.path.basename(name):
-        raise SuspiciousFileOperation("File name '%s' includes path elements" % name)
+        raise UnsafeFileOperation("File name '%s' includes path elements" % name)
 
     return name
