+++ ellar/core/routing/__init__.py
@@ -1,6 +1,7 @@
 from .app import ApplicationRouter
 from .builder import RouterBuilder, get_controller_builder_factory
 from .factory import ControllerRouterFactory
+from .file_mount import AppStaticFileMount, ASGIFileMount
 from .module_router import ModuleRouterFactory
 
 __all__ = [
@@ -9,4 +10,6 @@ __all__ = [
     "get_controller_builder_factory",
     "RouterBuilder",
     "ModuleRouterFactory",
+    "AppStaticFileMount",
+    "ASGIFileMount",
 ]
